pyproject.toml
[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[project]
name = "simplify-downloader"
version = "0.1.0"
description = "Headless Playwright downloader for TumbleDry MIS"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "playwright>=1.55.0",
    "python-dotenv>=1.0.0",
]

# Turn off packaging (we’re using Poetry for env + deps + scripts)
[tool.poetry]
package-mode = false
